requirements.txt
discord==2.2.3
PyMuPDF==1.22.3
Pillow==9.5.0
